Created MonsterController
Created the monster controller with all of its events.
Create NoiseDetectionManager as a non-object singleton with a unity event to help decouple noise events.
Created two test monster states.
Moved the monster's collider to the base object and added a rigidbody2d. (will need to move the sprite to a child object and make it counter rotate sot it does not rotate as the monster looks around)