Mimic Doesn't Change Layers Anymore
Mimic now is just set to Entities. I turned on IsTrigger during idle form regarding its Circle Collider 2D. So the player can walk on top of it like a regular item. When player interacts with it and the mimic switches state, the isTrigger becomes false so the player and it can collide. Player can now shoot projectiles to kill/injure the mimic during its idle state.