Updated Scavenger Wander Logic & Fixed Item Pick Up Bug
Scavenger Wander Logic should be less likely to get stuck on walls or in the same position (will transfer similar logic for the threatened state for an accurate path that is not invalid. Additionally, while testing, I caught an issue where the Scavenger attempts to pick up an item, does not but still sets the game object false, so I fixed that issue.